sword animation

+++ Assets/ScriptableObjects/Weapon.cs
@@ -6,7 +6,7 @@ public class Weapon : ScriptableObject
 {
     public int id;
     public int damage;
-    public float attackSpeed;
+    public int attackSpeed;
     public int durability;
     public float projectileSpeed;
     public Sprite image;
